refactor: move Table to components folder and split it int two components
